requirements.txt
discord.py
python-dotenv
PyQt6